Move try_print_query_stack to rustc_interface.
